requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
Django==5.1.1
django-allauth==64.2.0
gunicorn==23.0.0
h11==0.14.0
idna==3.8
iniconfig==2.0.0
joblib==1.4.2
jwt==1.3.1
numpy==2.1.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pluggy==1.5.0
pycparser==2.22
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
threadpoolctl==3.5.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
whitenoise==6.8.2
wsproto==1.2.0
